+++ src/ui/components/Timeline/Timeline.css
@@ -87,6 +87,9 @@
   cursor: pointer;
   padding: 4px;
   border-radius: 4px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .timeline .commands > button:hover,
